Optimize runtime caller capability checks
